+++ mitra_activitypub/src/importers.rs
@@ -339,7 +339,7 @@ pub fn is_actor_importer_error(error: &HandlerError) -> bool {
             HandlerError::DatabaseError(DatabaseError::NotFound(_)))
 }
 
-async fn perform_webfinger_query(
+pub(crate) async fn perform_webfinger_query(
     agent: &FederationAgent,
     actor_address: &ActorAddress,
 ) -> Result<String, HandlerError> {