+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add `'` and `;` to the list of stop characters for mentions and hashtags.
 - Don't allow HTML tags in display name.
 - Add missing `url` value to Mastodon API `Status` entities.
+- Determine webfinger address when importing portable actors.
 
 ### Fixed